backend/requirements.txt
fastapi==0.121.3
uvicorn[standard]==0.31.1
sqlalchemy==2.0.44
pydantic==2.12.4
pydantic-settings==2.12.0
python-jose==3.3.0
passlib[argon2]==1.7.4
python-multipart==0.0.6
email-validator>=2.0.0
cryptography>=41.0.0
argon2-cffi>=24.1.0
python-docx>=0.8.11
PyPDF2>=3.0.0
